fix and update logging and bump python
