+++ .github/workflows/ci.yml
@@ -11,7 +11,7 @@ on:
         default: "false"
         required: true
 env:
-  main-python-version: "3.12"
+  main-python-version: "3.13"
 
 jobs:
   lint:
@@ -38,10 +38,10 @@ jobs:
       max-parallel: 4
       matrix:
         python-version:
-          - "3.9"
           - "3.10"
           - "3.11"
           - "3.12"
+          - "3.13"
     steps:
       - uses: actions/checkout@v4
       - name: Install uv
@@ -77,7 +77,7 @@ jobs:
       matrix:
         os: [ubuntu-latest]
         python-version:
-          - "3.12"
+          - "3.13"
     needs: test
     steps:
       - uses: actions/checkout@v4
