+++ zeldarose/train_transformer.py
@@ -29,114 +29,7 @@ from pytorch_lightning import (
 
 from zeldarose.common import TrainConfig, TrainingModule
 from zeldarose.tasks import mbart, mlm, rtd
-from zeldarose.utils import dump_environment
-
-
-def setup_logging(
-    console: rich.console.Console,
-    verbose: bool,
-    log_file: Optional[pathlib.Path] = None,
-    replace_warnings: bool = True,
-):
-    logger.remove(0)  # Remove the default logger
-    if "SLURM_JOB_ID" in os.environ:
-        local_id = os.environ.get("SLURM_LOCALID", "someproc")
-        node_name = os.environ.get("SLURMD_NODENAME", "somenode")
-        appname = (
-            f"zeldarose ({os.environ.get('SLURM_PROCID', 'somerank')} [{local_id}@{node_name}])"
-        )
-    else:
-        appname = "zeldarose"
-
-    if verbose:
-        log_level = "DEBUG"
-        log_fmt = (
-            f"\\[{appname}]"
-            " [green]{time:YYYY-MM-DD HH:mm:ss.SSS}[/green] | [blue]{level: <8}[/blue] |"
-            " {message}"
-        )
-    else:
-        logging.getLogger(None).setLevel(logging.CRITICAL)
-        log_level = "INFO"
-        log_fmt = (
-            f"\\[{appname}]"
-            " [green]{time:YYYY-MM-DD}T{time:HH:mm:ss}[/green] {level} "
-            " {message}"
-        )
-
-    logger.add(
-        lambda m: console.print(m, end=""),
-        colorize=True,
-        format=log_fmt,
-        level=log_level,
-    )
-
-    if log_file:
-        logger.add(
-            log_file,
-            colorize=False,
-            format=(f"[{appname}] {{time:YYYY-MM-DD HH:mm:ss.SSS}} | {{level: <8}} | {{message}}"),
-            level="DEBUG",
-        )
-
-    # Deal with stdlib.logging
-
-    class InterceptHandler(logging.Handler):
-        def __init__(self, wrapped_name: Optional[str] = None, *args, **kwargs):
-            self.wrapped_name = wrapped_name
-            super().__init__(*args, **kwargs)
-
-        def emit(self, record):
-            # Get corresponding Loguru level if it exists
-            try:
-                level = logger.level(record.levelname).name
-            except ValueError:
-                level = record.levelno
-
-            # Find caller from where originated the logged message
-            frame, depth = logging.currentframe(), 2
-            while frame is not None and frame.f_code.co_filename == logging.__file__:
-                frame = frame.f_back
-                depth += 1
-
-            if frame is None:
-                warnings.warn(
-                    "Catching calls to logging is impossible in stackless environment,"
-                    " logging from external libraries might be lost.",
-                    stacklevel=2,
-                )
-            else:
-                if self.wrapped_name is not None:
-                    logger.opt(depth=depth, exception=record.exc_info).log(
-                        level, f"[bold]{self.wrapped_name} says:[/bold] {record.getMessage()}"
-                    )
-                else:
-                    logger.opt(depth=depth, exception=record.exc_info).log(
-                        level, record.getMessage()
-                    )
-
-    for libname in (
-        "datasets",
-        "huggingface_hub",
-        "lightning_fabric",
-        "pytorch_lightning",
-        "torch",
-        "torchmetrics",
-        "transformers",
-    ):
-        lib_logger = logging.getLogger(libname)
-        # FIXME: ugly, but is there a better way? What if they rely on having other handlers?
-        if lib_logger.handlers:
-            lib_logger.handlers = []
-        lib_logger.addHandler(InterceptHandler(libname))
-        logger.info(f"Intercepting logging from {libname}")
-
-    # Deal with stdlib.warnings
-    def showwarning(message, category, filename, lineno, file=None, line=None):
-        logger.warning(warnings.formatwarning(message, category, filename, lineno, None).strip())
-
-    if replace_warnings:
-        warnings.showwarning = showwarning
+from zeldarose.utils import dump_environment, setup_logging
 
 
 class SavePretrainedModelCallback(pl.Callback):
